+++ src/RcppExports.cpp
@@ -12,8 +12,8 @@ Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
 #endif
 
 // blv
-Rcpp::List blv(const arma::mat& costs, const arma::vec& X, double alpha, double beta, const arma::vec& Z_init, double epsilon, int iter_max, int conv_check, double precision, bool quadratic);
-RcppExport SEXP _blvim_blv(SEXP costsSEXP, SEXP XSEXP, SEXP alphaSEXP, SEXP betaSEXP, SEXP Z_initSEXP, SEXP epsilonSEXP, SEXP iter_maxSEXP, SEXP conv_checkSEXP, SEXP precisionSEXP, SEXP quadraticSEXP) {
+Rcpp::List blv(const arma::mat& costs, const arma::vec& X, double alpha, double beta, const arma::vec& Z_init, const arma::vec& kappa, double epsilon, int iter_max, int conv_check, double precision, bool quadratic, bool use_kappa);
+RcppExport SEXP _blvim_blv(SEXP costsSEXP, SEXP XSEXP, SEXP alphaSEXP, SEXP betaSEXP, SEXP Z_initSEXP, SEXP kappaSEXP, SEXP epsilonSEXP, SEXP iter_maxSEXP, SEXP conv_checkSEXP, SEXP precisionSEXP, SEXP quadraticSEXP, SEXP use_kappaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -22,12 +22,14 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< double >::type alpha(alphaSEXP);
     Rcpp::traits::input_parameter< double >::type beta(betaSEXP);
     Rcpp::traits::input_parameter< const arma::vec& >::type Z_init(Z_initSEXP);
+    Rcpp::traits::input_parameter< const arma::vec& >::type kappa(kappaSEXP);
     Rcpp::traits::input_parameter< double >::type epsilon(epsilonSEXP);
     Rcpp::traits::input_parameter< int >::type iter_max(iter_maxSEXP);
     Rcpp::traits::input_parameter< int >::type conv_check(conv_checkSEXP);
     Rcpp::traits::input_parameter< double >::type precision(precisionSEXP);
     Rcpp::traits::input_parameter< bool >::type quadratic(quadraticSEXP);
-    rcpp_result_gen = Rcpp::wrap(blv(costs, X, alpha, beta, Z_init, epsilon, iter_max, conv_check, precision, quadratic));
+    Rcpp::traits::input_parameter< bool >::type use_kappa(use_kappaSEXP);
+    rcpp_result_gen = Rcpp::wrap(blv(costs, X, alpha, beta, Z_init, kappa, epsilon, iter_max, conv_check, precision, quadratic, use_kappa));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -48,7 +50,7 @@ END_RCPP
 }
 
 static const R_CallMethodDef CallEntries[] = {
-    {"_blvim_blv", (DL_FUNC) &_blvim_blv, 10},
+    {"_blvim_blv", (DL_FUNC) &_blvim_blv, 12},
     {"_blvim_we_oc", (DL_FUNC) &_blvim_we_oc, 5},
     {NULL, NULL, 0}
 };