+++ NEWS.md
@@ -1,6 +1,15 @@
 # blvim (development version)
 
-# usblvim 0.1.1
+## New features
+
+* `blvim()` now supports conversion factors between attractivenesses and 
+  incoming flows, together with the new function `sim_conversion()`.
+
+## Minor improvements and bug fixes
+
+* improved test compatibility between different platforms
+
+# blvim 0.1.1
 * Simplify examples in `grid_autoplot()` to have them run in less than 5 
   seconds.